+++ core/src/plugins/filed/grpc/grpc.cc
@@ -132,7 +132,6 @@ struct plugin_ctx {
     return true;
   }
 
-
   bool setup(PluginContext* bareos_ctx, const void* data)
   {
     if (!bareos_ctx || !data) { return false; }
@@ -166,10 +165,69 @@ struct plugin_ctx {
   bool needs_setup() { return !child.has_value(); }
 
 
+  struct request_deleter {
+    void operator()(bareos::plugin::handlePluginEventRequest* req) const
+    {
+      delete_request(req);
+    }
+  };
+
+  using req_ptr = std::unique_ptr<bareos::plugin::handlePluginEventRequest,
+                                  request_deleter>;
+
+  static inline req_ptr make_event_request(filedaemon::bEvent* e, void* data)
+  {
+    return req_ptr{
+        to_grpc(static_cast<filedaemon::bEventType>(e->eventType), data)};
+  }
+
+  struct cached_event {
+    filedaemon::bEventType event_type;
+    req_ptr request;
+
+    filedaemon::bEventType type() const { return event_type; }
+    bareos::plugin::handlePluginEventRequest* req() const
+    {
+      return request.get();
+    }
+  };
+
+  bool cache_event(filedaemon::bEvent* e, void* data)
+  {
+    auto req = make_event_request(e, data);
+    if (!req) { return false; }
+    cached_events.emplace_back(cached_event{
+        static_cast<filedaemon::bEventType>(e->eventType), std::move(req)});
+
+    return true;
+  }
+
+  std::vector<cached_event> clear_cached_events()
+  {
+    std::vector<cached_event> events{};
+    std::swap(events, cached_events);
+    return events;
+  }
+
+
  public:
+  std::vector<cached_event> cached_events;
   std::string name;
   std::string cmd;
   std::string plugin_name;
+
+  struct c_free {
+    void operator()(void* ptr) const { free(ptr); }
+  };
+
+  template <typename T> using c_ptr = std::unique_ptr<T, c_free>;
+
+  std::size_t name_storage_size = 0;
+  c_ptr<char> name_storage;
+
+  std::size_t object_storage_size = 0;
+  c_ptr<char> object_storage;
+
 
   std::optional<grpc_child> child;
 };
@@ -187,6 +245,7 @@ bRC newPlugin(PluginContext* ctx)
   /* the actual setup is done inside of handle plugin event, because
    * at the moment we have no idea which plugin to start! */
 
+  // these are the events that this plugin _needs_ to work
   RegisterBareosEvent(ctx, filedaemon::bEventPluginCommand);
   RegisterBareosEvent(ctx, filedaemon::bEventNewPluginOptions);
   RegisterBareosEvent(ctx, filedaemon::bEventPluginCommand);
@@ -195,6 +254,28 @@ bRC newPlugin(PluginContext* ctx)
   RegisterBareosEvent(ctx, filedaemon::bEventEstimateCommand);
   RegisterBareosEvent(ctx, filedaemon::bEventBackupCommand);
   RegisterBareosEvent(ctx, filedaemon::bEventRestoreObject);
+  RegisterBareosEvent(ctx, filedaemon::bEventNewPluginOptions);
+
+  // these are the early events that the module _might_ want,
+  // but would not get since its initialised _after_ they are emitted
+  RegisterBareosEvent(ctx, filedaemon::bEventJobStart);
+  RegisterBareosEvent(ctx, filedaemon::bEventStartBackupJob);
+  RegisterBareosEvent(ctx, filedaemon::bEventStartRestoreJob);
+  RegisterBareosEvent(ctx, filedaemon::bEventStartVerifyJob);
+  RegisterBareosEvent(ctx, filedaemon::bEventLevel);
+  RegisterBareosEvent(ctx, filedaemon::bEventSince);
+  RegisterBareosEvent(ctx, filedaemon::bEventRestoreObject);
+  RegisterBareosEvent(ctx, filedaemon::bEventOptionPlugin);
+
+  RegisterBareosEvent(ctx, filedaemon::bEventVssInitializeForBackup);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssInitializeForRestore);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssSetBackupState);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssPrepareForBackup);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssBackupAddComponents);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssPrepareSnapshot);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssCreateSnapshots);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssRestoreLoadComponentMetadata);
+  RegisterBareosEvent(ctx, filedaemon::bEventVssRestoreSetComponentsSelected);
 
   return bRC_OK;
 }
@@ -239,9 +320,29 @@ bRC handlePluginEvent(PluginContext* ctx, filedaemon::bEvent* event, void* data)
 
       DebugLog(ctx, 100, FMT_STRING("using cmd string \"{}\" for the plugin"),
                plugin->cmd);
+
+      auto cached = plugin->clear_cached_events();
+
+      DebugLog(ctx, 100, FMT_STRING("inserting {} cached events"),
+               cached.size());
+
+      bool cached_err = false;
+      for (auto& cached_event : cached) {
+        DebugLog(ctx, 100, FMT_STRING("inserting cached {}-event"),
+                 static_cast<std::size_t>(cached_event.type()));
+        if (plugin->child->con.handlePluginEvent(cached_event.type(),
+                                                 cached_event.req())
+            == bRC_Error) {
+          cached_err = true;
+        }
+      }
+
+
       // we do not want to give "grpc:" to the plugin
-      return plugin->child->con.handlePluginEvent(bEventPluginCommand,
-                                                  (void*)plugin->cmd.c_str());
+      bRC res = plugin->child->con.handlePluginEvent(
+          bEventPluginCommand, (void*)plugin->cmd.c_str());
+      if (cached_err) { return bRC_Error; }
+      return res;
     } break;
     case bEventNewPluginOptions:
       [[fallthrough]];
@@ -260,39 +361,43 @@ bRC handlePluginEvent(PluginContext* ctx, filedaemon::bEvent* event, void* data)
                                                   (void*)plugin->cmd.c_str());
     } break;
     case bEventRestoreObject: {
-      if (data == nullptr) {
-        return plugin->child->con.handlePluginEvent(
-            bEventType(event->eventType), nullptr);
+      // the case where data == nullptr, happens if this is the last
+      // restore object for this plugin.   We can treat that specific
+      // event like any other event, as it requeres no special handling
+      // (no re_setup, etc.)
+      if (data != nullptr) {
+        auto* rop = reinterpret_cast<restore_object_pkt*>(data);
+        if (!plugin->re_setup(ctx, rop->plugin_name)) { return bRC_Error; }
+
+        char* old = rop->plugin_name;
+        rop->plugin_name = const_cast<char*>(plugin->cmd.c_str());
+        auto res = plugin->child->con.handlePluginEvent(
+            bEventType(event->eventType), (void*)rop);
+        rop->plugin_name = old;
+        return res;
       }
-
-      auto* rop = reinterpret_cast<restore_object_pkt*>(data);
-      if (!plugin->re_setup(ctx, rop->plugin_name)) { return bRC_Error; }
-
-      char* old = rop->plugin_name;
-      rop->plugin_name = const_cast<char*>(plugin->cmd.c_str());
-      auto res = plugin->child->con.handlePluginEvent(
-          bEventType(event->eventType), (void*)rop);
-      rop->plugin_name = old;
-      return res;
     } break;
+
     default: {
-      if (plugin->needs_setup()) {
-        DebugLog(
-            100,
-            FMT_STRING(
-                "cannot handle event {} as context was not set up properly"),
-            event->eventType);
-        return bRC_Error;
-      }
+      // intentionally left blank: normal events are handled below
     } break;
   }
 
-  if (!plugin->child) {
-    JobLog(ctx, M_FATAL, FMT_STRING("plugin is not running"));
-    return bRC_Error;
+  if (plugin->needs_setup()) {
+    DebugLog(100,
+             FMT_STRING("cannot handle event {} as context was not set up "
+                        "yet, caching ..."),
+             event->eventType);
+    if (!plugin->cache_event(event, data)) {
+      JobLog(ctx, M_FATAL, FMT_STRING("could not cache event {}"),
+             event->eventType);
+      return bRC_Error;
+    }
+    return bRC_OK;
+  } else {
+    return plugin->child->con.handlePluginEvent(
+        (filedaemon::bEventType)(event->eventType), data);
   }
-  return plugin->child->con.handlePluginEvent(
-      (filedaemon::bEventType)(event->eventType), data);
 }
 
 bRC startBackupFile(PluginContext* ctx, filedaemon::save_pkt* pkt)
@@ -445,3 +550,34 @@ extern "C" int loadPlugin(filedaemon::PluginApiDefinition* core_info,
 }
 
 extern "C" int unloadPlugin() { return 0; }
+
+char* get_name_storage(PluginContext* ctx, std::size_t minsize)
+{
+  auto pctx = get(ctx);
+
+  if (pctx->name_storage_size == 0) {
+    pctx->name_storage.reset(static_cast<char*>(malloc(minsize)));
+    pctx->name_storage_size = minsize;
+  } else if (pctx->name_storage_size < minsize) {
+    pctx->name_storage.reset(
+        static_cast<char*>(realloc(pctx->name_storage.release(), minsize)));
+    pctx->name_storage_size = minsize;
+  }
+
+  return pctx->name_storage.get();
+}
+char* get_object_storage(PluginContext* ctx, std::size_t minsize)
+{
+  auto pctx = get(ctx);
+
+  if (pctx->object_storage_size == 0) {
+    pctx->object_storage.reset(static_cast<char*>(malloc(minsize)));
+    pctx->object_storage_size = minsize;
+  } else if (pctx->object_storage_size < minsize) {
+    pctx->object_storage.reset(
+        static_cast<char*>(realloc(pctx->object_storage.release(), minsize)));
+    pctx->object_storage_size = minsize;
+  }
+
+  return pctx->object_storage.get();
+}