+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix next-pool overrides by job [PR #2279]
 - stored: fix race condition [PR #2359]
 - dird: fix nextvol crash [PR #2335]
+- grpc: fix missing some early events [PR #2333]
 
 ### Documentation
 - docs: fix grpc-fd plugin call [PR #2068]
@@ -206,6 +207,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #2328]: https://github.com/bareos/bareos/pull/2328
 [PR #2329]: https://github.com/bareos/bareos/pull/2329
 [PR #2332]: https://github.com/bareos/bareos/pull/2332
+[PR #2333]: https://github.com/bareos/bareos/pull/2333
 [PR #2335]: https://github.com/bareos/bareos/pull/2335
 [PR #2340]: https://github.com/bareos/bareos/pull/2340
 [PR #2341]: https://github.com/bareos/bareos/pull/2341